Extract mergeability rules to a separate trait
